+++ diagnosticism/internal/__init__.py
@@ -3,6 +3,47 @@ import os
 import platform
 
 
+TRUE_STRINGS    =   [
+    'true',
+    'True',
+    'TRUE',
+    'on',
+    'On',
+    'ON',
+    'yes',
+    'Yes',
+    'YES',
+    '1',
+]
+
+FALSE_STRINGS   =   [
+    'false',
+    'False',
+    'FALSE',
+    'off',
+    'Off',
+    'OFF',
+    'no',
+    'No',
+    'NO',
+    '0',
+]
+
+TRUE_STRINGS_lower  =   [
+    'true',
+    'on',
+    'yes',
+    '1',
+]
+
+FALSE_STRINGS_lower =   [
+    'false',
+    'off',
+    'no',
+    '0',
+]
+
+
 def _is_windows():
     """
     Obtains whether the operating system is a flavour of Windows.
@@ -103,3 +144,78 @@ def _basename(path):
 
     return path[ix_last_slash + 1:]
 
+
+def _str2bool(s, default_value=None):
+    """
+    Converts a string to a boolean, by consideration of the human-readable
+    intention of the string. For example,
+
+    Returns
+    -------
+    True | False | None
+    """
+
+    if s in TRUE_STRINGS:
+
+        return True
+
+
+    if s in FALSE_STRINGS:
+
+        return False
+
+
+    s = s.trim().lower()
+
+
+    if s in TRUE_STRINGS_lower:
+
+        return True
+
+
+    if s in FALSE_STRINGS_lower:
+
+        return False
+
+
+    return default_value
+
+
+def _bool_from_env(args, fn_name):
+
+    if len(args) == 1:
+
+        v = args[0]
+
+        if not isinstance(v, bool):
+
+            raise TypeError("`%s()` argument must be `True` or `False`" % fn_name)
+
+        return v
+
+    if len(args) == 2:
+
+        def_val =   args[1]
+
+        if isinstance(args[0], (list, tuple)):
+
+            ev_names    =   args[0]
+        else:
+
+            ev_names    =   [ args[0] ]
+
+        for ev_name in ev_names:
+
+            if ev_name in os.environ:
+
+                ev_val = os.environ[ev_name]
+
+                return _str2bool(ev_val, def_val)
+
+        if not isinstance(def_val, bool):
+
+            raise TypeError("`%s()` second argument must be `True` or `False`" % fn_name)
+
+        return def_val
+
+    raise TypeError("`%s()` takes 1 or 2 arguments" % fn_name)